% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/acpc_reorient.R
\name{acpc_reorient}
\alias{acpc_reorient}
\title{AC/PC Reorientation}
\usage{
acpc_reorient(infiles, spmdir = spm_dir(), verbose = TRUE)
}
\arguments{
\item{infiles}{(character) Files to reorient.  First file will be used to
estimate AC/PC, then rest will be transformed}

\item{spmdir}{(character) path for SPM8.  If NULL, assumes
SPM8 is in matlabpath and so is spm8/toolbox
Must have nii_setorigin installed.  In
\code{system.file("", package="cttools")} from
http://www.mccauslandcenter.sc.edu/CRNL/sw/spm8/spm.zip}

\item{verbose}{(logical) Print diagnostic output}
}
\value{
Exit code from MATLAB.  If not zero, there was an error
}
\description{
Function that AC/PC re-orients the images for SPM
spatial normalization routine.  Uses nii_setorigin from
http://www.mccauslandcenter.sc.edu/CRNL/sw/spm8/spm.zip
}

