% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_spm_dir.R
\name{add_spm_dir}
\alias{add_spm_dir}
\alias{add_spm12_dir}
\title{Add SPM Directory}
\usage{
add_spm_dir(x, spmdir = spm_dir(verbose = verbose), verbose = TRUE)

add_spm12_dir(x, spmdir = spm_dir(verbose = verbose), verbose = TRUE)
}
\arguments{
\item{x}{Character vector of commands}

\item{spmdir}{SPM Directory}

\item{verbose}{Print diagnostic messages}
}
\value{
A character vector
}
\description{
Add SPM Directory
}
