% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filename.R
\name{filename_check}
\alias{filename_check}
\title{Perform filename checks for SPM}
\usage{
filename_check(filename, ...)
}
\arguments{
\item{filename}{filename of an image or nifti object}

\item{...}{arguments passed to \code{\link{checknii}}}
}
\value{
Character of filename
}
\description{
Checks a filename to see if nifti and expands paths
to absolute paths
}
\seealso{
\code{\link{checknii}}
}

