% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_xyz.R
\name{group_xyz}
\alias{group_xyz}
\title{Center of Gravity for Multiple Areas}
\usage{
group_xyz(img, k = 1, ...)
}
\arguments{
\item{img}{Object of class nifti}

\item{k}{Minimum number of voxels for a cluster.
See \code{\link{spm_bwlabel}}}

\item{...}{Arguments passed to \code{\link{spm_bwlabel}}}
}
\value{
Matrix of 3 columns
}
\description{
Find Center of Gravity of Each Area of Image, after thresholding
}
