% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntime_points.R
\name{ntime_points}
\alias{ntime_points}
\title{Get number of time points from a file}
\usage{
ntime_points(filename)
}
\arguments{
\item{filename}{List of nifti objects, a vector of character 
filenames, or a single 4D nifti}
}
\value{
Vector of time points
}
\description{
Extracts the number of time points from a nifti 
object, list or character
}
