% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spm12_script.R
\name{run_spm12_script}
\alias{run_spm12_script}
\title{Wrapper for running \code{spm12_script}}
\usage{
run_spm12_script(script_name, jobvec = NULL, mvec = NULL,
  add_spm_dir = TRUE, spmdir = spm_dir(), clean = TRUE, verbose = TRUE,
  ...)
}
\arguments{
\item{script_name}{Name of the script filename without .m ext,
passed to \code{\link{spm12_script}}}

\item{jobvec}{Vector of characters to be substituted in _job.m file}

\item{mvec}{Vector of characters to be substituted in .m file}

\item{add_spm_dir}{Add SPM12 directory from this package}

\item{spmdir}{SPM dir to add, will use package default directory}

\item{clean}{Remove scripts from temporary directory after running}

\item{verbose}{Print diagnostic messages}

\item{...}{Arguments to pass to \code{\link{spm12_script}}}
}
\value{
Result of \code{\link{run_matlab_script}}
}
\description{
Runs \code{\link{spm12_script}} with wrapper for 
spm12r functions
}
