% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm12_condition.R
\name{spm12_condition}
\alias{spm12_condition}
\title{Title}
\usage{
spm12_condition(name, onset, duration, time_mod_order = 0, param_mod = NULL,
  orth = TRUE)
}
\arguments{
\item{name}{}

\item{onset}{}

\item{duration}{}

\item{time_mod_order}{}

\item{param_mod}{}

\item{orth}{}

\item{cond}{List of conditions}
}
\value{
A list of objects, each with a \code{name}, \code{onset}, 
\code{duration}, and other condition values.
}
\description{
Title
}
\examples{
res = spm12_condition(
name = "condition1", 
onset = c(0, 2, 4, 6, 8),
duration = rep(1, 5) )
print(res)
L = list(
cond1 = list(onset = c(0, 2, 4, 6, 8), duration = rep(1, 5)), 
cond2 = list(onset = c(0, 2, 4, 6, 8) + 1, duration = rep(1, 5))
)
res = spm12_condition_list(L)
print(res) 
}
