% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm12_contrast_manager.R
\name{spm12_contrast_manager}
\alias{spm12_contrast_manager}
\alias{build_spm12_contrast_manager}
\title{SPM12 fMRI Contrasts}
\usage{
spm12_contrast_manager(..., add_spm_dir = TRUE, spmdir = spm_dir(verbose =
  verbose), clean = TRUE, verbose = TRUE)

build_spm12_contrast_manager(spm, delete_existing = TRUE,
  contrast_list = NULL, ...)
}
\arguments{
\item{...}{Arguments passed to 
\code{\link{matlabbatch_to_script}}}

\item{add_spm_dir}{Add SPM12 directory from this package}

\item{spmdir}{SPM dir to add, will use package default directory}

\item{clean}{Remove scripts from temporary directory after running}

\item{verbose}{Print diagnostic messages}

\item{spm}{Path to SPM.mat file}

\item{write_residuals}{Should residuals be written?}

\item{method}{Method for model estimation}

\item{bayesian}{If method = "Bayesian", this is for a 1st level
model Bayesian estimation and this list specifies the 
parameters}

\item{...}{Arguments passed to \code{\link{run_spm12_script}}}
}
\value{
A list of output and results
}
\description{
SPM12 fMRI Contrasts
}
