% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm12_coregister.R
\name{spm12_coregister.deprecated}
\alias{spm12_coregister.deprecated}
\title{SPM12 Coregister (Estimate and Reslice)}
\usage{
spm12_coregister.deprecated(
  fixed,
  moving,
  other.files = NULL,
  prefix = "r",
  add_spm_dir = TRUE,
  spmdir = spm_dir(),
  clean = TRUE,
  verbose = TRUE,
  outdir = NULL,
  install_dir = NULL,
  ...
)
}
\arguments{
\item{fixed}{File that is assumed fixed}

\item{moving}{moving file to be registered to fixed space}

\item{other.files}{Other files to register to fixed, in same 
space as moving file}

\item{prefix}{Prefix to append to front of image filename}

\item{add_spm_dir}{Add SPM12 directory from this package}

\item{spmdir}{SPM dir to add, will use package default directory}

\item{clean}{Remove scripts from temporary directory after running}

\item{verbose}{Print diagnostic messages}

\item{outdir}{Directory to copy results.  If full filename given, 
then results will
be in \code{dirname(filename)}}

\item{install_dir}{directory to download SPM12}

\item{...}{Arguments passed to \code{\link{run_spm12_script}}}
}
\value{
Result from run_matlab_script
}
\description{
Performs SPM12 coregistration estimation and 
reslicing on an Image
}
