% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm12_fmri_first_level.R
\name{build_spm12_first_level}
\alias{build_spm12_first_level}
\alias{spm12_first_level}
\title{SPM12 fMRI First Level Model}
\usage{
build_spm12_first_level(..., outdir = NULL, est_args = list(write_residuals
  = FALSE, method = "Classical", bayesian = NULL), verbose = TRUE)

spm12_first_level(..., outdir = NULL, est_args = list(write_residuals =
  FALSE, method = "Classical", bayesian = NULL), add_spm_dir = TRUE,
  spmdir = spm_dir(verbose = verbose), clean = TRUE, verbose = TRUE)
}
\arguments{
\item{...}{Arguments passed to 
\code{\link{build_spm12_first_level_spec}}}

\item{outdir}{output directory for results}

\item{est_args}{Arguments passed to 
\code{\link{build_spm12_fmri_est}}}

\item{verbose}{Print diagnostic messages}

\item{add_spm_dir}{Add SPM12 directory from this package}

\item{spmdir}{SPM dir to add, will use package default directory}

\item{clean}{Remove scripts from temporary directory 
after running}
}
\value{
A list of objects, including an spm object 
and output files.
}
\description{
SPM12 fMRI First Level Model
}
