% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spm12_normalize_est.R
\name{spm12_normalize_est}
\alias{spm12_normalize_est}
\title{SPM12 Normalize (Estimate)}
\usage{
spm12_normalize_est(filename, add_spm_dir = TRUE, spmdir = spm_dir(),
  clean = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{filename}{File to be normalized to the template}

\item{add_spm_dir}{Add SPM12 directory from this package}

\item{spmdir}{SPM dir to add, will use package default directory}

\item{clean}{Remove scripts from temporary directory after running}

\item{verbose}{Print diagnostic messages}

\item{...}{Arguments passed to \code{\link{run_spm12_script}}}
}
\value{
Result from run_matlab_script
}
\description{
Estimate SPM12 (Spatial) Normalization from image
}

