% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm12_normalize_write.R
\name{spm12_normalize_write}
\alias{spm12_normalize_write}
\title{SPM12 Normalize (Write)}
\usage{
spm12_normalize_write(filename, other.files = NULL, retimg = TRUE,
  reorient = FALSE, add_spm_dir = TRUE, spmdir = spm_dir(),
  clean = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{filename}{Filename of deformation (nifti)}

\item{other.files}{Files to be written using the estimated
normalization}

\item{retimg}{Logical indicating if image should be returned or
result from \code{\link{run_matlab_script}}}

\item{reorient}{if \code{retimg=TRUE} pass to \code{\link{readNIfTI}}}

\item{add_spm_dir}{Add SPM12 directory from this package}

\item{spmdir}{SPM dir to add, will use package default directory}

\item{clean}{Remove scripts from temporary directory after running}

\item{verbose}{Print diagnostic messages}

\item{...}{Arguments passed to \code{\link{run_spm12_script}}}
}
\value{
Result from run_matlab_script
}
\description{
Applies SPM12 (Spatial) Normalization to images
}

