% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm12_realign.R, R/zzz_spm12_realign.R
\name{spm12_realign}
\alias{spm12_realign}
\alias{spm12_realign}
\title{SPM12 Realign (Estimate and Reslice)}
\usage{
spm12_realign(filename, fwhm = 5, quality = 0.9, separation = 4,
  register_to = c("first", "mean"), est_interp = c("bspline2", "trilinear",
  paste0("bspline", 3:7)), wrap_x = FALSE, wrap_y = FALSE, wrap_z = FALSE,
  weight_image = NULL, reslice = c("all+mean", "all", "2:n", "mean"),
  reslice_interp = c("bspline4", "nearestneighbor", "trilinear",
  paste0("bspline", 2:3), paste0("bspline", 5:7), "fourier"), mask = FALSE,
  prefix = "r", add_spm_dir = TRUE, spmdir = spm_dir(verbose = verbose),
  clean = TRUE, verbose = TRUE, outdir = NULL, execute = TRUE, ...)

spm12_realign(filename, fwhm = 5, quality = 0.9, separation = 4,
  register_to = c("first", "mean"), est_interp = c("bspline2", "trilinear",
  paste0("bspline", 3:7)), wrap_x = FALSE, wrap_y = FALSE, wrap_z = FALSE,
  weight_image = NULL, reslice = c("all+mean", "all", "2:n", "mean"),
  reslice_interp = c("bspline4", "nearestneighbor", "trilinear",
  paste0("bspline", 2:3), paste0("bspline", 5:7), "fourier"), mask = FALSE,
  prefix = "r", add_spm_dir = TRUE, spmdir = spm_dir(verbose = verbose),
  clean = TRUE, verbose = TRUE, outdir = NULL, execute = TRUE, ...)
}
\arguments{
\item{filename}{Files to be realigned and resliced}

\item{fwhm}{Full-Width Half Max to smooth}

\item{quality}{Quality versus speed trade-off.
Highest quality (1) gives most precise results, 
whereas lower qualities gives faster realignment.}

\item{separation}{The  average  distance  between  sampled points (in mm).
Can be a vector to allow a coarse registration followed by increasingly fine}

\item{register_to}{Should the files be registered to the first or the mean}

\item{est_interp}{Interpolator for estimation}

\item{wrap_x}{wrap in x-direction}

\item{wrap_y}{wrap in y-direction}

\item{wrap_z}{wrap in z-direction}

\item{weight_image}{weighting image to weight each voxel of the reference 
image during estimation. The weights are proportional to the 
inverses of the standard deviations. May be used when there is
a lot of motion.}

\item{reslice}{Options for reslicing all - all images in filename,
2:n - all images in filename 2:length(filename),
all+mean - all images and the mean, mean - mean only}

\item{reslice_interp}{Interpolator for reslicing}

\item{mask}{Mask the data.  With masking enabled, the program searches 
through the whole time series looking for voxels which need to be sampled 
from outside  the  original  images.  Where  this  occurs, 
 that  voxel is set to zero for the whole set of images}

\item{prefix}{Prefix to append to front of image filename}

\item{add_spm_dir}{Add SPM12 directory from this package}

\item{spmdir}{SPM dir to add, will use package default directory}

\item{clean}{Remove scripts from temporary directory after running}

\item{verbose}{Print diagnostic messages}

\item{outdir}{Directory to copy results.  If full filename given, then results will
be in \code{dirname(filename)}}

\item{execute}{Should the script be executed or just return
the  \code{matlabbatch}  object}

\item{...}{Arguments passed to \code{\link{run_spm12_script}}}

\item{filename}{Files to be realigned and resliced}

\item{fwhm}{Full-Width Half Max to smooth}

\item{register_to}{Should the files be registered to the first or the mean}

\item{reslice}{Options for reslicing all - all images in filename,
2:n - all images in filename 2:length(filename),
all+mean - all images and the mean, mean - mean only}

\item{prefix}{Prefix to append to front of image filename}

\item{add_spm_dir}{Add SPM12 directory from this package}

\item{spmdir}{SPM dir to add, will use package default directory}

\item{clean}{Remove scripts from temporary directory after running}

\item{verbose}{Print diagnostic messages}

\item{outdir}{Directory to copy results.  If full filename given, then results will
be in \code{dirname(filename)}}

\item{fwhm}{Gaussian  smoothing  to  apply  to  the 256x256 joint histogram. 
Other information theoretic coregistration methods use fewer bins,}

\item{...}{Arguments passed to \code{\link{run_spm12_script}}}
}
\value{
Result from run_matlab_script

List of output files, the \code{matlabbatch} object, and the script
}
\description{
Performs SPM12 realignment estimation and reslicing on an Image

Performs SPM12 realignment estimation and reslicing on an Image
}
