% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm12_results.R
\name{spm12_results}
\alias{spm12_results}
\alias{build_spm12_results}
\title{SPM12 Results Report}
\usage{
spm12_results(..., add_spm_dir = TRUE, spmdir = spm_dir(verbose = verbose),
  clean = TRUE, verbose = TRUE)

build_spm12_results(spm, units = c("Volumetric", "Scalp-Time",
  "Scalp-Frequency", "Time-Frequency", "Frequency-Frequency"),
  result_format = c("none", "ps", "eps", "pdf", "jpg", "tif", "fig", "csv",
  "nidm"), write_images = c("none", "threshold_spm", "binary_clusters",
  "nary_clusters"), contrast_list = NULL, image_basename = NULL, ...)
}
\arguments{
\item{...}{Arguments passed to 
\code{\link{matlabbatch_to_script}}}

\item{add_spm_dir}{Add SPM12 directory from this package}

\item{spmdir}{SPM dir to add, will use package default directory}

\item{clean}{Remove scripts from temporary directory after running}

\item{verbose}{Print diagnostic messages}

\item{spm}{Path to SPM.mat file}

\item{write_residuals}{Should residuals be written?}

\item{method}{Method for model estimation}

\item{bayesian}{If method = "Bayesian", this is for a 1st level
model Bayesian estimation and this list specifies the 
parameters}

\item{...}{Arguments passed to \code{\link{run_spm12_script}}}
}
\value{
A list of output and results
}
\description{
SPM12 Results Report
}
