% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spm12_script.R
\name{spm12_script}
\alias{spm12_script}
\title{Find SPM12 Script}
\usage{
spm12_script(script_name, outdir = tempdir())
}
\arguments{
\item{script_name}{Name of the script filename without ".m" ext}

\item{outdir}{Path to copy scripts and run}
}
\value{
Chracter vector of script paths
}
\description{
Copies the SPM12 script from the scripts directory
to a temporary file
}
\examples{
spm12_script(script_name = "Segment")
}

