% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spm12_slice_timing.R
\name{spm12_slice_timing}
\alias{spm12_slice_timing}
\title{SPM12 Slice Timing Correction}
\usage{
spm12_slice_timing(filename, nslices, tr, ta = tr - tr/nslices,
  slice_order = 1:nslices, ref_slice = slice_order[median(seq(nslices))],
  prefix = "a", add_spm_dir = TRUE, spmdir = spm_dir(), clean = TRUE,
  verbose = TRUE, ...)
}
\arguments{
\item{filename}{Files to be slice-time corrected}

\item{nslices}{Number of slices in the images}

\item{tr}{Repetition time (in seconds)}

\item{ta}{Time between the first and the last slice within one scan}

\item{slice_order}{Order slices were taken (if not specified, assumed ascending),
bottom slice = 1}

\item{ref_slice}{Reference slice}

\item{prefix}{Prefix to append to front of image filename}

\item{add_spm_dir}{Add SPM12 directory from this package}

\item{spmdir}{SPM dir to add, will use package default directory}

\item{clean}{Remove scripts from temporary directory after running}

\item{verbose}{Print diagnostic messages}

\item{...}{Arguments passed to \code{\link{run_spm12_script}}}
}
\value{
Result from run_matlab_script
}
\description{
Performs SPM12 slice timing correction on images
}

