% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spm_probs_to_seg.R
\name{spm_probs_to_seg}
\alias{spm_probs_to_seg}
\title{SPM Probability to Hard Segmenation}
\usage{
spm_probs_to_seg(img, ties.method = c("first", "last", "random"))
}
\arguments{
\item{img}{list of images for probabilities for each class}

\item{ties.method}{a character string specifying how ties are handled.
See \code{\link{max.col}}.  Note, order of ties is different than
\code{\link{max.col}}.}
}
\value{
Object of class nifti
}
\description{
Converts probability images from SPM
segmentation to a hard, choose-one segmentation
}
\examples{
\dontrun{
spm_seg = spm12_segment(image)
seg = spm_probs_to_seg(spm_seg)
}
}

