% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_version.R
\name{spm_version}
\alias{spm_version}
\alias{spm12_version}
\title{SPM Version}
\usage{
spm_version()

spm12_version()
}
\value{
Character vector of length 1
}
\description{
SPM Version
}
\examples{
if (matlabr::have_matlab()) {
spm_version()
}
}
