% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_matlabbatch.R
\name{run_matlabbatch}
\alias{run_matlabbatch}
\alias{matlabbatch_job}
\alias{matlabbatch_to_script}
\title{Run Matlab Batch from List}
\usage{
run_matlabbatch(spm, add_spm_dir = TRUE, clean = TRUE, verbose = TRUE,
  gui = FALSE, ...)

matlabbatch_job(spm, add_spm_dir = TRUE, gui = FALSE, ...)

matlabbatch_to_script(spm, batch_prefix = "matlabbatch{1}.", ...)
}
\arguments{
\item{spm}{List of the class \code{matlabbatch}}

\item{add_spm_dir}{should SPM12 directory be added to the script}

\item{clean}{Remove scripts from temporary directory after running}

\item{verbose}{Print diagnostic messages}

\item{gui}{is the GUI necessary?}

\item{...}{additional arguments to pass to 
\code{\link{matlabbatch_job}}}

\item{batch_prefix}{prefix to add to the names of \code{spm}}
}
\value{
Result of \code{\link{run_matlab_script}}
}
\description{
Run Matlab Batch from List
}
