% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm12_contrast_manager.R
\name{spm12_contrast_manager}
\alias{spm12_contrast_manager}
\alias{build_spm12_contrast_manager}
\title{SPM12 fMRI Contrasts}
\usage{
spm12_contrast_manager(
  ...,
  add_spm_dir = TRUE,
  spmdir = spm_dir(verbose = verbose),
  clean = TRUE,
  verbose = TRUE,
  install_dir = NULL
)

build_spm12_contrast_manager(
  spm,
  delete_existing = TRUE,
  contrast_list = NULL,
  ...
)
}
\arguments{
\item{...}{Arguments passed to 
\code{\link{matlabbatch_to_script}}}

\item{add_spm_dir}{Add SPM12 directory from this package}

\item{spmdir}{SPM dir to add, will use package default directory}

\item{clean}{Remove scripts from temporary directory after running}

\item{verbose}{Print diagnostic messages}

\item{install_dir}{directory to download SPM12}

\item{spm}{Path to SPM.mat file}

\item{delete_existing}{Delete existing contrasts}

\item{contrast_list}{List of contrasts to pass to 
\code{\link{spm12_contrast_list}} for conversion}
}
\value{
A list of output and results
}
\description{
SPM12 fMRI Contrasts
}
