% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spm12_realign.R
\name{spm12_realign}
\alias{spm12_realign}
\title{SPM12 Realign (Estimate)}
\usage{
spm12_realign(filename, fwhm = 5, register_to = c("first", "mean"),
  prefix = "r", add_spm_dir = TRUE, spmdir = spm_dir(), clean = TRUE,
  verbose = TRUE, outdir = NULL, retimg = TRUE, reorient = FALSE, ...)
}
\arguments{
\item{filename}{Files to be realigned}

\item{fwhm}{Full-Width Half Max to smooth}

\item{register_to}{Should the files be registered to the first or the mean}

\item{prefix}{Prefix to append to front of image filename}

\item{add_spm_dir}{Add SPM12 directory from this package}

\item{spmdir}{SPM dir to add, will use package default directory}

\item{clean}{Remove scripts from temporary directory after running}

\item{verbose}{Print diagnostic messages}

\item{outdir}{Directory to copy results.  If full filename given, then results will
be in \code{dirname(filename)}}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readNIfTI}}.}

\item{...}{Arguments passed to \code{\link{run_spm12_script}}}
}
\value{
Result from run_matlab_script
}
\description{
Performs SPM12 Realignment estimation on an Image
}

