% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/spm12_segment.R
\name{spm12_segment}
\alias{spm12_segment}
\title{SPM12 Segment}
\usage{
spm12_segment(filename, add_spm_dir = TRUE)
}
\arguments{
\item{filename}{File to be segmented}

\item{outdir}{Output directory}
}
\value{
Result from run_matlab_script
}
\description{
Performs SPM12 Segmentation on an Image
}

