% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/spm12_segment.R
\name{spm12_segment}
\alias{spm12_segment}
\title{SPM12 Segment}
\usage{
spm12_segment(filename, reorient = TRUE, add_spm_dir = TRUE, clean = TRUE,
  verbose = TRUE)
}
\arguments{
\item{filename}{File to be segmented}

\item{reorient}{Run \code{\link{acpc_reorient}} on image first.
Warning, this will set the orientation differently}

\item{add_spm_dir}{Add SPM12 directory from this package}

\item{clean}{Remove scripts from temporary directory after running}

\item{verbose}{Print diagnostic messages}
}
\value{
Result from run_matlab_script
}
\description{
Performs SPM12 Segmentation on an Image
}

