% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spm12_segment.R
\name{spm12_segment}
\alias{spm12_segment}
\title{SPM12 Segment}
\usage{
spm12_segment(filename, retimg = TRUE, set_origin = TRUE,
  add_spm_dir = TRUE, spmdir = spm_dir(), clean = TRUE, verbose = TRUE,
  reorient = FALSE, ...)
}
\arguments{
\item{filename}{File to be segmented}

\item{retimg}{Logical indicating if image should be returned or
result from \code{\link{run_matlab_script}}}

\item{set_origin}{Run \code{\link{acpc_reorient}} on image first.
Warning, this will set the orientation differently}

\item{add_spm_dir}{Add SPM12 directory from this package}

\item{spmdir}{SPM dir to add, will use package default directory}

\item{clean}{Remove scripts from temporary directory after running}

\item{verbose}{Print diagnostic messages}

\item{reorient}{if \code{retimg=TRUE} pass to \code{\link{readNIfTI}}}

\item{...}{Arguments passed to \code{\link{run_spm12_script}}}
}
\value{
Result from run_matlab_script or nifti file, depending on
\code{retimg}
}
\description{
Performs SPM12 Segmentation on an Image
}

