% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_bwlabel.R
\name{spm_bwlabel}
\alias{spm_bwlabel}
\title{SPM BWLabel Clusters of Certain Size}
\usage{
spm_bwlabel(
  infile,
  outfile = NULL,
  retimg = TRUE,
  k = NULL,
  topN = NULL,
  margin = NULL,
  binary = TRUE,
  spmdir = spm_dir(verbose = verbose, install_dir = install_dir),
  reorient = FALSE,
  verbose = TRUE,
  install_dir = NULL
)
}
\arguments{
\item{infile}{input filename}

\item{outfile}{output filename}

\item{retimg}{Return the image instead of matlab output}

\item{k}{Minimum cluster size needed}

\item{topN}{Top number of clusters kept (used if k is \code{NULL})}

\item{margin}{Margin to loop over if wanted in 2D}

\item{binary}{(logical) Should the result be binary or numbered with cluster.}

\item{spmdir}{SPM directory (for MATLAB)}

\item{reorient}{If \code{retimg}, then this argument is passed to 
\code{readNIfTI}}

\item{verbose}{Print Diagnostics}

\item{install_dir}{directory to download SPM12}
}
\value{
Output from \code{run_matlab_script} or \code{nifti} object,
depending on \code{retimg}

Result from \code{\link{run_matlab_script}}
}
\description{
Get Cluster of certain size from spm_bwlabel
}
\note{
Taken from 
http://en.wikibooks.org/wiki/SPM/How-to#How_to_remove_clusters_under_a_certain_size_in_a_binary_mask.3F
}
\examples{
if (matlabr::have_matlab()) { 
d = rep(25, 3)
arr = array(rbinom(prod(d), size = 1, prob = 0.5), dim = d)
nim = oro.nifti::nifti(arr)
res = spm_bwlabel(nim, k = 1, binary = FALSE, install_dir = tempdir())
tab = table(c(res))
tab
out_arr = as(res, "array")
}
}
