% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_dir.R
\name{spm_dir}
\alias{spm_dir}
\title{Get SPM12 Directory}
\usage{
spm_dir(verbose = FALSE)
}
\arguments{
\item{verbose}{print diagnostic messages, passed to 
\code{\link{install_spm12}}}
}
\value{
Character vector of spm12 paths
}
\description{
Returns the SPM12 directory
}
